pyproject.toml
[project]
name = "inventory-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "aiogram>=3.22.0",
    "asyncio>=4.0.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.31",
    "langgraph>=0.6.6",
    "psycopg>=3.2.9",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
